+++ mayan/apps/document_states/views/workflow_views.py
@@ -30,15 +30,17 @@ from ..forms import (
 )
 from ..icons import (
     icon_workflow_list, icon_workflow_state, icon_workflow_state_action,
-    icon_workflow_transition
+    icon_workflow_transition, icon_workflow_transition_field
 )
 from ..links import (
     link_setup_workflow_create, link_setup_workflow_state_create,
     link_setup_workflow_state_action_selection,
-    link_setup_workflow_transition_create
+    link_setup_workflow_transition_create,
+    link_setup_workflow_transition_field_create,
 )
 from ..models import (
-    Workflow, WorkflowState, WorkflowStateAction, WorkflowTransition
+    Workflow, WorkflowState, WorkflowStateAction, WorkflowTransition,
+    WorkflowTransitionField
 )
 from ..permissions import (
     permission_workflow_create, permission_workflow_delete,
@@ -729,6 +731,124 @@ class SetupWorkflowTransitionTriggerEventListView(ExternalObjectMixin, FormView)
             viewname='document_states:setup_workflow_transition_list',
             kwargs={'pk': self.get_object().workflow.pk}
         )
+
+
+# Transition fields
+
+class SetupWorkflowTransitionFieldCreateView(ExternalObjectMixin, SingleObjectCreateView):
+    external_object_class = WorkflowTransition
+    external_object_permission = permission_workflow_edit
+    fields = (
+        'name', 'label', 'field_type', 'help_text', 'required', 'widget',
+        'widget_kwargs'
+    )
+    def get_extra_context(self):
+        return {
+            'navigation_object_list': ('transition', 'workflow'),
+            'transition': self.external_object,
+            'title': _(
+                'Create a field for workflow transition: %s'
+            ) % self.external_object,
+            'workflow': self.external_object.workflow
+        }
+
+    def get_instance_extra_data(self):
+        return {
+            'transition': self.external_object,
+        }
+
+    def get_queryset(self):
+        return self.external_object.fields.all()
+
+    def get_post_action_redirect(self):
+        return reverse(
+            viewname='document_states:setup_workflow_transition_field_list',
+            kwargs={'pk': self.external_object.pk}
+        )
+
+
+class SetupWorkflowTransitionFieldDeleteView(SingleObjectDeleteView):
+    model = WorkflowTransitionField
+    object_permission = permission_workflow_edit
+
+    def get_extra_context(self):
+        return {
+            'navigation_object_list': (
+                'object', 'workflow_transition', 'workflow'
+            ),
+            'object': self.object,
+            'title': _('Delete workflow transition field: %s') % self.object,
+            'workflow': self.object.transition.workflow,
+            'workflow_transition': self.object.transition,
+        }
+
+    def get_post_action_redirect(self):
+        return reverse(
+            viewname='document_states:setup_workflow_transition_field_list',
+            kwargs={'pk': self.object.transition.pk}
+        )
+
+
+class SetupWorkflowTransitionFieldEditView(SingleObjectEditView):
+    fields = (
+        'name', 'label', 'field_type', 'help_text', 'required', 'widget',
+        'widget_kwargs'
+    )
+    model = WorkflowTransitionField
+    object_permission = permission_workflow_edit
+
+    def get_extra_context(self):
+        return {
+            'navigation_object_list': (
+                'object', 'workflow_transition', 'workflow'
+            ),
+            'object': self.object,
+            'title': _('Edit workflow transition field: %s') % self.object,
+            'workflow': self.object.transition.workflow,
+            'workflow_transition': self.object.transition,
+        }
+
+    def get_post_action_redirect(self):
+        return reverse(
+            viewname='document_states:setup_workflow_transition_field_list',
+            kwargs={'pk': self.object.transition.pk}
+        )
+
+
+class SetupWorkflowTransitionFieldListView(ExternalObjectMixin, SingleObjectListView):
+    external_object_class = WorkflowTransition
+    external_object_permission = permission_workflow_edit
+
+    def get_extra_context(self):
+        return {
+            'hide_object': True,
+            'navigation_object_list': ('object', 'workflow'),
+            'no_results_icon': icon_workflow_transition_field,
+            'no_results_main_link': link_setup_workflow_transition_field_create.resolve(
+                context=RequestContext(
+                    request=self.request, dict_={
+                        'object': self.external_object
+                    }
+                )
+            ),
+            'no_results_text': _(
+                'Workflow transition fields allow adding data to the '
+                'workflow\'s context. This additional context data can then '
+                'be used by other elements of the workflow system like the '
+                'workflow state actions.'
+            ),
+            'no_results_title': _(
+                'There are no fields for this workflow transition'
+            ),
+            'object': self.external_object,
+            'title': _(
+                'Fields for workflow transition: %s'
+            ) % self.external_object,
+            'workflow': self.external_object.workflow,
+        }
+
+    def get_source_queryset(self):
+        return self.external_object.fields.all()
 
 
 class ToolLaunchAllWorkflows(ConfirmView):